restructure the package without `a` prefix
