--- django_async_extensions/aforms/__init__.py
@@ -1,4 +1,0 @@
-from django_async_extensions.aforms.models import AsyncModelForm
-
-
-__all__ = ("AsyncModelForm",)
