+++ django_async_extensions/forms/__init__.py
@@ -1,0 +1,4 @@
+from django_async_extensions.forms.models import AsyncModelForm
+
+
+__all__ = ("AsyncModelForm",)
